backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.3.26
langchain-core >=0.3
langchain-text-splitters==0.3.8
langchain-groq==0.2.0   # Adjust this version as needed
reportlab==4.0.7
pydantic==2.11.7